requirements.txt
numpy>=1.24.0
numpy-stl>=3.0.1
Pillow>=10.0.0
matplotlib>=3.7.0
vtk>=9.2.0
python-dotenv>=1.0.0
# Hugging Face depth-estimation support
transformers>=4.41.0
torch>=2.2.0
timm>=0.9.0